# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Property Management System API
  description: API for managing properties, leases, user authentication, and payments for landlords and tenants. Tested endpoints are marked as [TESTED].
  version: 1.0.2
  contact:
    name: PMS Support
    email: support@pms-bd.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://pms-bd.onrender.com
    description: Production server
paths:
...
  /api/payments:
    post:
      summary: Create a payment [TESTED]
      tags: [Payments]
      description: Initiates a payment for a lease (tenant only, mock payment with UUID). Tested with lease ID 6835edfa996e720d12ea30d0.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            example:
              lease_id: 6835edfa996e720d12ea30d0
              amount: 560000
      responses:
        '201':
          description: Payment processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  payment: { $ref: '#/components/schemas/Payment' }
              example:
                message: Payment processed
                payment:
                  _id: new_payment_id
                  lease_id: 6835edfa996e720d12ea30d0
                  tenant_id: 6835ae50eea8975c7a3a130d
                  amount: 560000
                  status: completed
                  transaction_id: uuid_string
                  created_at: 2025-05-27T16:53:14.145Z
        '400':
          description: Missing lease_id or amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized (non-tenant or invalid lease)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List payments [TESTED]
      tags: [Payments]
      description: Returns payments (tenant sees own, landlord sees for their properties). Tested with tenant token.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
              example:
                - _id: new_payment_id
                  lease_id:
                    _id: 6835edfa996e720d12ea30d0
                    start_date: 2025-06-01T00:00:00.000Z
                    end_date: 2026-05-31T00:00:00.000Z
                  tenant_id: 6835ae50eea8975c7a3a130d
                  amount: 560000
                  status: completed
                  transaction_id: uuid_string
                  created_at: 2025-05-27T16:53:14.145Z
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/payments/subaccount:
    post:
      summary: Create payment subaccount
      tags: [Payments]
      description: Sets up a payment subaccount for landlord (e.g., Paystack). Untested, assumed implemented.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [businessName, bankCode, accountNumber]
              properties:
                businessName: { type: string }
                bankCode: { type: string }
                accountNumber: { type: string }
            example:
              businessName: Landlord Business
              bankCode: 044
              accountNumber: 0690000031
      responses:
        '201':
          description: Subaccount created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  subaccountId: { type: string }
              example:
                message: Subaccount created
                subaccountId: sub_123456
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/payments/initialize:
    post:
      summary: Initialize a payment
      tags: [Payments]
      description: Initiates a payment (e.g., rent payment, Paystack). Untested, assumed implemented.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, lease_id]
              properties:
                amount: { type: number }
                lease_id: { type: string }
            example:
              amount: 560000
              lease_id: 6835edfa996e720d12ea30d0
      responses:
        '200':
          description: Payment initialized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  paymentUrl: { type: string }
              example:
                message: Payment initialized
                paymentUrl: https://paystack.com/pay/123456
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    PaymentRequest:
      type: object
      required: [lease_id, amount]
      properties:
        lease_id: { type: string }
        amount: { type: number }
    Payment:
      type: object
      properties:
        _id: { type: string }
        lease_id:
          oneOf:
            - type: string
            - type: object
              properties:
                _id: { type: string }
                start_date: { type: string, format: date-time }
                end_date: { type: string, format: date-time }
        tenant_id: { type: string }
        amount: { type: number }
        status: { type: string, enum: [pending, completed, failed] }
        transaction_id: { type: string }
        created_at: { type: string, format: date-time }
    Error:
      type: object
      properties:
        message: { type: string }
      example:
        message: Invalid input
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []